requirements.txt
PyAutoGUI==0.9.54
discord.py==2.3.2
psutil==5.9.8
asyncio==3.4.3
pyinstaller==6.4.0